Updated packages migration

+++ src/afp2/database/migrations/2021_03_01_203909_create_packages_table.php
@@ -15,8 +15,8 @@ class CreatePackagesTable extends Migration
     {
         Schema::create('packages', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('product_id');
-            $table->integer('order_id');
+            $table->unsignedBigInteger('product_id');
+            $table->unsignedBigInteger('order_id');
             $table->integer('amount');
             $table->foreign('product_id')->references('id')->on('products');
             $table->foreign('order_id')->references('id')->on('orders');
